Move the weekday name fill into @woocommerce/date
The fill lived in the admin client, so only pages loading the wc-admin
app or embed bundle got it. @woocommerce/date already owns the moment
locale corrections (ensureMomentStartOfWeek), and both calendar
components import from it, so putting it there covers every consumer.
