+++ plugins/woocommerce/client/admin/client/embed.tsx
@@ -2,11 +2,8 @@
  * Internal dependencies
  */
 import { initRemoteLogging } from './lib/init-remote-logging';
-import { initDateLocale } from './lib/date-locale';
 // Initialize remote logging early to log any errors that occur during initialization.
 initRemoteLogging();
-// Fill in the translated weekday names moment is missing, before anything renders a calendar.
-initDateLocale();
 
 /**
  * Internal dependencies
