+++ plugins/woocommerce/client/admin/client/index.tsx
@@ -11,11 +11,8 @@ import {
  * Internal dependencies
  */
 import { initRemoteLogging } from './lib/init-remote-logging';
-import { initDateLocale } from './lib/date-locale';
 // Initialize remote logging early to log any errors that occur during initialization.
 initRemoteLogging();
-// Fill in the translated weekday names moment is missing, before anything renders a calendar.
-initDateLocale();
 
 /**
  * Internal dependencies
